Safari-611.1.10/LayoutTests/http/tests/navigation/redirect-to-fragment.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpResourceLoadCallbacks();
    testRunner.waitUntilDone();
}

onload = function() {
    if (window.testRunner)
        testRunner.notifyDone();
}
</script>
</head>
<body>
<p>Test passes if WebKit ignores the request fragment identifier after the redirect to a url with a fragment identifier.</p>
<iframe src="resources/redirect-to-fragment.php#foo"></iframe>
</body>
</html>
